requirements.txt
Django==3.2.*
djangorestframework==3.14.*
drf-yasg==1.21.*
gunicorn==21.2.*
Pillow==10.1.*
psycopg2-binary==2.9.*
python-dotenv==1.0.*
stripe==7.10.*
django-active-link==0.1.*